ref(onboarding): Split node awslambda onboarding docs
