+++ static/app/gettingStartedDocs/node/awslambda/agentMonitoring.tsx
@@ -1,0 +1,5 @@
+import {getNodeAgentMonitoringOnboarding} from 'sentry/utils/gettingStartedDocs/node';
+
+export const agentMonitoring = getNodeAgentMonitoringOnboarding({
+  packageName: '@sentry/aws-serverless',
+});
