+++ static/app/gettingStartedDocs/node/awslambda/index.tsx
@@ -1,0 +1,21 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+
+import {agentMonitoring} from './agentMonitoring';
+import {crashReport} from './crashReport';
+import {logs} from './logs';
+import {mcp} from './mcp';
+import {onboarding} from './onboarding';
+import {profiling} from './profiling';
+import {platformOptions, type PlatformOptions} from './utils';
+
+const docs: Docs<PlatformOptions> = {
+  onboarding,
+  crashReportOnboarding: crashReport,
+  profilingOnboarding: profiling,
+  logsOnboarding: logs,
+  agentMonitoringOnboarding: agentMonitoring,
+  mcpOnboarding: mcp,
+  platformOptions,
+};
+
+export default docs;
