+++ static/app/gettingStartedDocs/node/awslambda/logs.tsx
@@ -1,0 +1,6 @@
+import {getNodeLogsOnboarding} from 'sentry/utils/gettingStartedDocs/node';
+
+export const logs = getNodeLogsOnboarding({
+  docsPlatform: 'aws-lambda',
+  packageName: '@sentry/aws-serverless',
+});
