+++ static/app/gettingStartedDocs/node/awslambda/mcp.tsx
@@ -1,0 +1,5 @@
+import {getNodeMcpOnboarding} from 'sentry/utils/gettingStartedDocs/node';
+
+export const mcp = getNodeMcpOnboarding({
+  packageName: '@sentry/aws-serverless',
+});
