+++ static/app/gettingStartedDocs/node/awslambda/onboarding.spec.tsx
@@ -4,7 +4,8 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs, {InstallationMethod} from './awslambda';
+import {InstallationMethod} from './utils';
+import docs from '.';
 
 describe('awslambda onboarding docs', () => {
   describe('Lambda Layer', () => {
