+++ static/app/gettingStartedDocs/node/awslambda/profiling.tsx
@@ -1,0 +1,5 @@
+import {getNodeProfilingOnboarding} from 'sentry/utils/gettingStartedDocs/node';
+
+export const profiling = getNodeProfilingOnboarding({
+  packageName: '@sentry/aws-serverless',
+});
